scripts/quant-analytics/requirements.txt
FinanceToolkit>=1.4.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pandas>=2.2.0
